Address CodeRabbit review on deep linking, settings, and tray E2E specs.
Use distinct channel preconditions for running deep links, shared settings
window helper, resilient server map polling, mainWindow fixture, tray helpers,
and accurate DL-02 test naming.
