+++ e2e/specs/deep_linking/oauth_callback.test.ts
@@ -2,7 +2,6 @@
 // See LICENSE.txt for license information.
 
 import {test, expect} from '../../fixtures/index';
-import {waitForAppReady} from '../../helpers/appReadiness';
 import {demoConfig} from '../../helpers/config';
 import {mattermostDeepLinkUrl, openDeepLinkInApp} from '../../helpers/deeplink';
 import {buildServerMap} from '../../helpers/serverMap';
@@ -10,9 +9,7 @@ import {buildServerMap} from '../../helpers/serverMap';
 test(
     'DL-03 OAuth callback deep link navigates the active server view',
     {tag: ['@P1', '@all']},
-    async ({electronApp}) => {
-        await waitForAppReady(electronApp);
-
+    async ({electronApp, mainWindow}) => {
         const serverName = demoConfig.servers[0].name;
         const oauthPath = '/oauth/authorize?client_id=desktop&response_type=code&state=e2e-test';
         const deepLink = mattermostDeepLinkUrl(`example.com${oauthPath}`);
@@ -28,10 +25,8 @@ test(
             message: 'OAuth callback deep link should navigate the example server view',
         }).toContain('example.com/oauth/authorize');
 
-        const mainWindow = electronApp.windows().find((window) => window.url().includes('index'));
-        expect(mainWindow).toBeDefined();
         await expect.poll(
-            () => mainWindow!.innerText('.ServerDropdownButton'),
+            () => mainWindow.innerText('.ServerDropdownButton'),
             {timeout: 15_000},
         ).toBe(serverName);
     },
